fix(bagheri, tmsaddon): small refactor on packaging. CustomException httpStatus code deleted.
