--- src/main/java/de/ampada/tmsaddon/aspect/security/JwtTokenFilterConfigurer.java
@@ -1,22 +1,0 @@
-package de.ampada.tmsaddon.aspect.security;
-
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.DefaultSecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-public class JwtTokenFilterConfigurer extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-
-	private final JwtTokenProvider jwtTokenProvider;
-
-	public JwtTokenFilterConfigurer(JwtTokenProvider jwtTokenProvider) {
-		this.jwtTokenProvider = jwtTokenProvider;
-	}
-
-	@Override
-	public void configure(HttpSecurity http) {
-		JwtTokenFilter customFilter = new JwtTokenFilter(jwtTokenProvider);
-		http.addFilterBefore(customFilter, UsernamePasswordAuthenticationFilter.class);
-	}
-
-}
