+++ src/main/java/de/ampada/tmsaddon/configuration/jwt/JwtTokenFilter.java
@@ -1,6 +1,7 @@
-package de.ampada.tmsaddon.aspect.security;
+package de.ampada.tmsaddon.configuration.jwt;
 
 import de.ampada.tmsaddon.exception.CustomException;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
@@ -30,7 +31,7 @@ public class JwtTokenFilter extends OncePerRequestFilter {
         } catch (CustomException ex) {
             /*BAGHERI : this is very important, since it guarantees the user is not authenticated at all*/
             SecurityContextHolder.clearContext();
-            httpServletResponse.sendError(ex.getHttpStatus().value(), ex.getMessage());
+            httpServletResponse.sendError(HttpStatus.GONE.value(), ex.getMessage());
             return;
         }
