+++ src/main/java/de/ampada/tmsaddon/domains/Role.java
@@ -1,7 +1,11 @@
-package de.ampada.tmsaddon.entity;
+package de.ampada.tmsaddon.domains;
 
 public class Role {
     private String name;
+
+    public Role(String name) {
+        this.name = name;
+    }
 
     public String getName() {
         return name;
