+++ src/main/java/de/ampada/tmsaddon/exception/CustomException.java
@@ -1,27 +1,11 @@
 package de.ampada.tmsaddon.exception;
 
-import org.springframework.http.HttpStatus;
-
 public class CustomException extends RuntimeException {
 
     private static final long serialVersionUID = 1L;
 
-    private final String message;
-    private final HttpStatus httpStatus;
-
-    public CustomException(String message, HttpStatus httpStatus) {
-        this.message = message;
-        this.httpStatus = httpStatus;
+    public CustomException(String message) {
+        super(message);
     }
-
-    @Override
-    public String getMessage() {
-        return message;
-    }
-
-    public HttpStatus getHttpStatus() {
-        return httpStatus;
-    }
-
 }
